requirements.txt
openai
tiktoken
aiofiles
nest_asyncio
python-dotenv
chardet
pyyaml
